Add unit tests for can message sending
The goal is for the user to force the usb board to send a CAN message.
The syntax for doing so is "MSSS,DD,DD,DD,DD,DD,DD,DD,DD;", where SSS is
the standard ID of the message (stored in can_msg_t.sid), and DD are
data bytes. Note that all characters are nibbles, or hex digits (numbers
0-9 and A-F).

The user does not have to supply all 8 data bytes, because CAN messages
can have fewer than that, but they can't supply more than 8. They're
allowed to type just one nibble instead of two or three (so if they want
to send a message with SID 0x11 and no data bytes, they can write
"M11;", and it should just work.

I'd also like to make all the characters case agnostic. So "M7AA;" is the
same as "m7aa;" is the same as "M7aA;". I haven't added tests to the
other commands ("G", "L", and "S") to make sure they're case agnostic,
but they should also be made that way in due course.

I've ran user_config.c through astyle to make it the same as all the
other files in the project, so if the diff is really big, that's why